Revamped how resources and states are handled
Enum identifiers for states and resources have been removed.
States are now referenced by a string name, and resources have better
automatic handling.

States must be registered with the game before game.run() is called.
Resources must be requested from a state, preferably in the constructor
of a state. The game engine will keep track of what resources are
necessary based on if a state has requested it.